% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{even_spaced_curves_impl}
\alias{even_spaced_curves_impl}
\title{R wrapper around the C++ implementation for even spaced curves.}
\usage{
even_spaced_curves_impl(
  x_start1,
  y_start1,
  n_curves1,
  n_steps1,
  min_steps_allowed1,
  step_length1,
  d_sep1,
  flow_field1,
  flow_field_width1
)
}
\arguments{
\item{x_start1}{the x coordinate of the starting point from which the function will start to draw your curve.}

\item{y_start1}{the y coordinate of the starting point from which the function will start to draw your curve.}

\item{n_curves1}{the number of curves you want to draw.}

\item{n_steps1}{the number of steps used to draw each curve.}

\item{min_steps_allowed1}{the minimum number of steps allowed in each curve (see Details for more info).}

\item{step_length1}{the length/distance taken in each step.}

\item{d_sep1}{the "separation distance", i.e., the amount of distance that each curve must be from neighbouring curves.}

\item{flow_field1}{a 2D matrix with double values, each double value represents an angle value.}

\item{flow_field_width1}{the width of the input flow field.}
}
\description{
Use the function \code{rlefer::even_spaced_curves()} instead.
}
\keyword{internal}
