% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{non_overlapping_curves_impl}
\alias{non_overlapping_curves_impl}
\title{R wrapper around the C++ implementation for non overlapping curves.}
\usage{
non_overlapping_curves_impl(
  starting_points,
  n_curves,
  n_steps,
  min_steps_allowed,
  step_length,
  d_sep,
  flow_field,
  flow_field_width
)
}
\arguments{
\item{starting_points}{a list object, with the x and y coordinates of the starting points for each curve.}

\item{n_curves}{the number of curves you want to draw.}

\item{n_steps}{the number of steps used to draw each curve.}

\item{min_steps_allowed}{the minimum number of steps allowed in each curve (see Details for more info).}

\item{step_length}{the length/distance taken in each step.}

\item{d_sep}{the "separation distance", i.e., the amount of distance that each curve must be from neighbouring curves.}

\item{flow_field}{a 2D matrix with double values, each double value represents an angle value.}

\item{flow_field_width}{the width of the input flow field.}
}
\description{
Use the function \code{rlefer::non_overlapping_curves()} instead.
}
\keyword{internal}
