% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_curves.R
\name{non_overlapping_curves}
\alias{non_overlapping_curves}
\title{Draws multiple non-overlapping curves in a flow field.}
\usage{
non_overlapping_curves(
  starting_points,
  n_steps,
  min_steps_allowed,
  step_length,
  d_sep,
  flow_field
)
}
\arguments{
\item{starting_points}{a list object, with the x and y coordinates of the starting points for each curve.}

\item{n_steps}{the number of steps used to draw each curve.}

\item{min_steps_allowed}{the minimum number of steps allowed in each curve (see Details for more info).}

\item{step_length}{the length/distance taken in each step.}

\item{d_sep}{the "separation distance", i.e., the amount of distance that each curve must be from neighbouring curves.}

\item{flow_field}{a 2D matrix with double values, each double value represents an angle value.}
}
\value{
This function returns a \code{tibble} object with 6 columns:
\itemize{
\item{\code{curve_id}: the ID of the curve.}
\item{\code{x}: the x coordinates of each point that represents the curve.}
\item{\code{y}: the y coordinates of each point that represents the curve.}
\item{\code{direction_id}: which direction that the algorithm was following when drawing the current point (0 means from left to right, 1 means from right to left).}
\item{\code{step_id}: the ID (or the number) of the current step.}
\item{\code{steps_taken}: the number of steps taken to draw the current curve.}
}
}
\description{
Draws multiple non-overlapping curves in a flow field,
using the Jobard and Lefer (1997) algorithm.
}
\examples{
library(ambient)
set.seed(50)
flow_field <- noise_perlin(c(240, 240))

}
\references{
Jobard, Bruno, and Wilfrid Lefer. 1997. “Creating Evenly-Spaced Streamlines of Arbitrary Density.” In Visualization in Scientific Computing ’97, edited by Wilfrid Lefer and Michel Grave, 43–55. Vienna: Springer Vienna.
}
