% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_curves.R
\name{point_order}
\alias{point_order}
\title{Get order ID of each point in the curve}
\usage{
point_order(direction_id)
}
\description{
Helper function to get the order ID of each point
in the curve.
}
\keyword{internal}
